+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### v0.9.3 | In development
+### v0.9.3 | 2017-05-29
 
 * Added support for Kwant v1.3.x and improved `Model.tokwant()` exporting of multi-orbital models.